+++ crm_backend/src/main/scala/agents/domain/entity/Agent.scala
@@ -1,0 +1,11 @@
+package agents.domain.entity
+
+case class Agent (
+  id: Long = -1,
+  documentType: String,
+  personType: String,
+  name: String,
+  lastName: String,
+  phone: String,
+  email:String
+)
