--- crm_backend/src/main/scala/authentication/application/create_user/RequestCreateUser.scala
@@ -1,12 +1,0 @@
-package authentication.application.create_user
-
-case class RequestCreateUser (
-  document: Long,
-  documentType: String,
-  name: String,
-  lastName: String,
-  phone: String,
-  email: String,
-  userName: String,
-  password: String
-)
