--- crm_backend/src/main/scala/authentication/application/create_user/ResponseCreateUser.scala
@@ -1,8 +1,0 @@
-package authentication.application.create_user
-
-import authentication.domain.entity.User
-
-case class ResponseCreateUser (
-  user: User,
-  agent: String
-)
