--- crm_backend/src/main/scala/authentication/application/login_user/RequestLoginUser.scala
@@ -1,6 +1,0 @@
-package authentication.application.login_user
-
-case class RequestLoginUser (
-  username: String,
-  password: String
-)
