--- crm_backend/src/main/scala/authentication/application/login_user/ResponseLoginUser.scala
@@ -1,5 +1,0 @@
-package authentication.application.login_user
-
-case class ResponseLoginUser (
-  jwt: String
-)
