--- crm_backend/src/main/scala/authentication/domain/entity/User.scala
@@ -1,9 +1,0 @@
-package authentication.domain.entity
-
-case class User (
-  id:Long,
-  username: String,
-  password: String,
-  description: Option[String],
-  agentId: Long
-)
