--- crm_backend/src/main/scala/authentication/domain/entity/UserRole.scala
@@ -1,6 +1,0 @@
-package authentication.domain.entity
-
-case class UserRole (
-  roleId: Long,
-  userId: Long
-)
