--- crm_backend/src/main/scala/authentication/domain/repository/AuthenticationRepository.scala
@@ -1,7 +1,0 @@
-package authentication.domain.repository
-
-import authentication.domain.entity.User
-
-trait AuthenticationRepository:
-  def getUserByUsername(username:String):Option[User]
-  def createUser(user:User):User
