--- crm_backend/src/main/scala/authentication/domain/service/JwtService.scala
@@ -1,4 +1,0 @@
-package authentication.domain.service
-
-trait JwtService:
-  def encodeUserInfo(username: String, userId:Long): String
