+++ crm_backend/src/main/scala/authentications/application/create_user/RequestCreateUser.scala
@@ -1,0 +1,12 @@
+package authentications.application.create_user
+
+case class RequestCreateUser (
+  document: Long,
+  documentType: String,
+  name: String,
+  lastName: String,
+  phone: String,
+  email: String,
+  userName: String,
+  password: String
+)
