+++ crm_backend/src/main/scala/authentications/application/create_user/ResponseCreateUser.scala
@@ -1,0 +1,8 @@
+package authentications.application.create_user
+
+import authentications.domain.entity.User
+
+case class ResponseCreateUser (
+  user: User,
+  agent: String
+)
