+++ crm_backend/src/main/scala/authentications/application/login_user/RequestLoginUser.scala
@@ -1,0 +1,6 @@
+package authentications.application.login_user
+
+case class RequestLoginUser (
+  username: String,
+  password: String
+)
