+++ crm_backend/src/main/scala/authentications/application/login_user/ResponseLoginUser.scala
@@ -1,0 +1,5 @@
+package authentications.application.login_user
+
+case class ResponseLoginUser (
+  jwt: String
+)
