+++ crm_backend/src/main/scala/authentications/domain/entity/User.scala
@@ -1,0 +1,9 @@
+package authentications.domain.entity
+
+case class User (
+  id:Long,
+  username: String,
+  password: String,
+  description: Option[String],
+  agentId: Long
+)
