+++ crm_backend/src/main/scala/authentications/domain/entity/UserRole.scala
@@ -1,0 +1,6 @@
+package authentications.domain.entity
+
+case class UserRole (
+  roleId: Long,
+  userId: Long
+)
