+++ crm_backend/src/main/scala/authentications/domain/repository/AuthenticationRepository.scala
@@ -1,0 +1,7 @@
+package authentications.domain.repository
+
+import authentications.domain.entity.User
+
+trait AuthenticationRepository:
+  def getUserByUsername(username:String):Option[User]
+  def createUser(user:User):User
