+++ crm_backend/src/main/scala/authentications/domain/service/JwtService.scala
@@ -1,0 +1,4 @@
+package authentications.domain.service
+
+trait JwtService:
+  def encodeUserInfo(username: String, userId:Long): String
