--- crm_backend/src/main/scala/authorization/domain/entity/Permission.scala
@@ -1,8 +1,0 @@
-package authorization.domain.entity
-
-case class Permission (
-  id: Long,
-  name: String,
-  accessModule: String,
-  roleId: Long
-)
