--- crm_backend/src/main/scala/authorization/domain/entity/Role.scala
@@ -1,7 +1,0 @@
-package authorization.domain.entity
-
-case class Role (
-  id: Long,
-  name: String,
-  description: Option[String]
-)
