+++ crm_backend/src/main/scala/authorizations/domain/entity/Permission.scala
@@ -1,0 +1,8 @@
+package authorizations.domain.entity
+
+case class Permission (
+  id: Long,
+  name: String,
+  accessModule: String,
+  roleId: Long
+)
