+++ crm_backend/src/main/scala/authorizations/domain/entity/Role.scala
@@ -1,0 +1,7 @@
+package authorizations.domain.entity
+
+case class Role (
+  id: Long,
+  name: String,
+  description: Option[String]
+)
